requirements.txt
grpcio
grpcio-tools
Flask
Flask-Cors
